Prototype
A very rough first prototype of the game. I only made guns and trap
doors for the mouse player to click on; They're marked in red. The
keyboard player tries to get to the green orb I've marked as the goal.
The mouse player tries to kill him.

I used a sprite from my older game, you can check to see how sprites
work through it.